+++ .dotnet/src/Custom/Assistants/DeleteThreadResult.cs
@@ -1,0 +1,10 @@
+﻿namespace OpenAI.Assistants;
+
+[CodeGenModel("DeleteThreadResponse")]
+public partial class DeleteThreadResult
+{
+    // CUSTOM: Made internal.
+    /// <summary> The object type, which is always `thread.deleted`. </summary>
+    [CodeGenMember("Object")]
+    internal InternalDeleteThreadResponseObject Object { get; } = InternalDeleteThreadResponseObject.ThreadDeleted;
+}