+++ .dotnet/src/Custom/Models/DeleteModelResult.cs
@@ -1,0 +1,10 @@
+﻿namespace OpenAI.Models;
+
+[CodeGenModel("DeleteModelResponse")]
+public partial class DeleteModelResult
+{
+    // CUSTOM: Made internal.
+    /// <summary> The object type, which is always `model`. </summary>
+    [CodeGenMember("Object")]
+    internal InternalDeleteModelResponseObject Object { get; } = InternalDeleteModelResponseObject.Model;
+}