+++ .dotnet/src/Custom/VectorStores/DeleteVectorStoreResult.cs
@@ -1,0 +1,9 @@
+﻿namespace OpenAI.VectorStores;
+
+[CodeGenModel("DeleteVectorStoreResponse")]
+public partial class DeleteVectorStoreResult
+{
+    // CUSTOM: Made internal.
+    /// <summary> The object type, which is always `vector_store.deleted`. </summary>
+    internal InternalDeleteVectorStoreResponseObject Object { get; } = InternalDeleteVectorStoreResponseObject.VectorStoreDeleted;
+}